+++ 1w-tdd/src/main/java/io/hhplus/tdd/point/dto/UserPoint.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.dto;
 
 public record UserPoint(
         long id,
